move search-related constants/actions/reducer into PdfSearch dir; simplify scrollToSearchTerm, PdfPage.componentDidUpdate
