+++ client/app/reader/PdfSearch/PdfSearchActions.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { createSearchAction } from 'redux-search';
 
-import * as Constants from '../constants';
+import * as Constants from './actionTypes';
 
 export const getDocumentText = (pdfDocument, file) =>
   (dispatch) => {
