+++ client/app/reader/PdfSearch/actionTypes.js
@@ -1,0 +1,6 @@
+export const GET_DOCUMENT_TEXT = 'GET_DCOUMENT_TEXT';
+export const UPDATE_SEARCH_INDEX = 'UPDATE_SEARCH_INDEX';
+export const SET_SEARCH_INDEX = 'SET_SEARCH_INDEX';
+export const SET_SEARCH_INDEX_TO_HIGHLIGHT = 'SET_SEARCH_INDEX_TO_HIGHLIGHT';
+export const UPDATE_SEARCH_INDEX_PAGE_INDEX = 'UPDATE_SEARCH_INDEX_PAGE_INDEX';
+export const UPDATE_SEARCH_RELATIVE_INDEX = 'UPDATE_SEARCH_RELATIVE_INDEX';
