+++ client/app/reader/constants.js
@@ -22,12 +22,6 @@ export const SET_PAGE_COORD_BOUNDS = 'SET_PAGE_COORD_BOUNDS';
 export const SET_UP_PAGE_DIMENSIONS = 'SET_UP_PAGE_DIMENSIONS';
 export const SET_PDF_DOCUMENT = 'SET_PDF_DOCUMENT';
 export const CLEAR_PDF_DOCUMENT = 'CLEAR_PDF_DOCUMENT';
-export const GET_DOCUMENT_TEXT = 'GET_DCOUMENT_TEXT';
-export const UPDATE_SEARCH_INDEX = 'UPDATE_SEARCH_INDEX';
-export const SET_SEARCH_INDEX = 'SET_SEARCH_INDEX';
-export const SET_SEARCH_INDEX_TO_HIGHLIGHT = 'SET_SEARCH_INDEX_TO_HIGHLIGHT';
-export const UPDATE_SEARCH_INDEX_PAGE_INDEX = 'UPDATE_SEARCH_INDEX_PAGE_INDEX';
-export const UPDATE_SEARCH_RELATIVE_INDEX = 'UPDATE_SEARCH_RELATIVE_INDEX';
 export const SET_DOCUMENT_LOAD_ERROR = 'SET_DOCUMENT_LOAD_ERROR';
 export const CLEAR_DOCUMENT_LOAD_ERROR = 'CLEAR_DOCUMENT_LOAD_ERROR';
 export const SET_LOADED_APPEAL_ID = 'SET_LOADED_APPEAL_ID';
